src/jump_host/server/requirements.txt
connexion[aiohttp,swagger-ui] >= 2.6.0; python_version>="3.6"
# 2.3 is the last version that supports python 3.5
connexion[aiohttp,swagger-ui] <= 2.3.0; python_version=="3.5" or python_version=="3.4"
# connexion requires werkzeug but connexion < 2.4.0 does not install werkzeug
# we must peg werkzeug versions below to fix connexion
# https://github.com/zalando/connexion/pull/1044
werkzeug == 0.16.1; python_version=="3.5" or python_version=="3.4"
swagger-ui-bundle == 0.0.9
aiohttp_jinja2 == 1.5.0
aiohttp_cors >= 0.7.0
